--- egs/voxforge/asr1/conf/train_k2.yaml
@@ -1,48 +1,0 @@
-# network architecture
-# encoder related
-transformer-input-layer: conv2d
-elayers: 12
-eunits: 2048
-# decoder related
-dlayers: 6
-dunits: 2048
-# attention related
-adim: 256
-aheads: 4
-# transformer related
-model-module: "espnet.nets.pytorch_backend.e2e_asr_conformer:E2E"
-
-# hybrid CTC/attention
-mtlalpha: 0.3
-
-# label smoothing
-lsm-weight: 0.1
-
-# minibatch related
-batch-size: 32
-maxlen-in: 512  # if input length  > maxlen_in, batchsize is automatically reduced
-maxlen-out: 150 # if output length > maxlen_out, batchsize is automatically reduced
-
-# optimization related
-sortagrad: 0 # Feed samples from shortest to longest ; -1: enabled for all epochs, 0: disabled, other: enabled for 'other' epochs
-opt: noam
-epochs: 200
-dropout-rate: 0.1
-accum-grad: 4
-grad-clip: 5
-patience: 0
-transformer-lr: 10.0
-transformer-warmup-steps: 25000
-transformer-attn-dropout-rate: 0.0
-transformer-length-normalized-loss: False
-transformer-init: pytorch
-
-# conformer specific setting
-transformer-encoder-pos-enc-layer-type: rel_pos
-transformer-encoder-selfattn-layer-type: rel_selfattn
-transformer-encoder-activation-type: swish
-macaron-style: true
-use-cnn-module: true
-cnn-module-kernel: 7
-
-ctc_type: k2
