--- egs/voxforge/asr1/train_model.sh
@@ -1,144 +1,0 @@
-#!/bin/bash
-
-# Copyright 2017 Johns Hopkins University (Shinji Watanabe)
-#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
-
-. ./path.sh || exit 1;
-. ./cmd.sh || exit 1;
-
-# general configuration
-backend=pytorch
-stage=-1       # start from -1 if you need to start from data download
-stop_stage=100
-ngpu=1         # number of gpus ("0" uses cpu, otherwise use gpu)
-debugmode=1
-dumpdir=dump   # directory to dump full features
-N=0            # number of minibatches to be used (mainly for debugging). "0" uses all minibatches.
-verbose=0      # verbose option
-resume=        # Resume the training from snapshot
-
-# feature configuration
-do_delta=false
-
-# config files
-preprocess_config=conf/no_preprocess.yaml  # use conf/specaug.yaml for data augmentation
-train_config=conf/train.yaml
-decode_config=conf/decode.yaml
-
-# decoding parameter
-recog_model=model.acc.best # set a model to be used for decoding: 'model.acc.best' or 'model.loss.best'
-
-# model average realted (only for transformer)
-n_average=10                 # the number of ASR models to be averaged
-use_valbest_average=false    # if true, the validation `n_average`-best ASR models will be averaged.
-                             # if false, the last `n_average` ASR models will be averaged.
-
-# data
-voxforge=downloads # original data directory to be stored
-lang=it # de, en, es, fr, it, nl, pt, ru
-
-# exp tag
-tag="" # tag for managing experiments.
-
-. utils/parse_options.sh || exit 1;
-
-# Set bash to 'debug' mode, it will exit on :
-# -e 'error', -u 'undefined variable', -o ... 'error in pipeline', -x 'print commands',
-set -e
-set -u
-set -o pipefail
-
-train_set=tr_${lang}
-train_dev=dt_${lang}
-recog_set="dt_${lang} et_${lang}"
-
-
-feat_tr_dir=${dumpdir}/${train_set}/delta${do_delta}; mkdir -p ${feat_tr_dir}
-feat_dt_dir=${dumpdir}/${train_dev}/delta${do_delta}; mkdir -p ${feat_dt_dir}
-dict=data/lang_1char/tr_${lang}_units.txt
-echo "dictionary: ${dict}"
-
-if [ -z ${tag} ]; then
-    expname=${train_set}_${backend}_$(basename ${train_config%.*})
-    if ${do_delta}; then
-        expname=${expname}_delta
-    fi
-else
-    expname=${train_set}_${backend}_${tag}
-fi
-expdir=exp/${expname}
-mkdir -p ${expdir}
-
-if [ ${stage} -le 3 ] && [ ${stop_stage} -ge 3 ]; then
-    echo "stage 3: Network Training"
-    ${cuda_cmd} --gpu ${ngpu} ${expdir}/train.log \
-        asr_train.py \
-        --config ${train_config} \
-        --preprocess-conf ${preprocess_config} \
-        --ngpu ${ngpu} \
-        --backend ${backend} \
-        --outdir ${expdir}/results \
-        --tensorboard-dir tensorboard/${expname} \
-        --debugmode ${debugmode} \
-        --dict ${dict} \
-        --debugdir ${expdir} \
-        --minibatches ${N} \
-        --verbose ${verbose} \
-        --resume ${resume} \
-        --train-json ${feat_tr_dir}/data.json \
-        --valid-json ${feat_dt_dir}/data.json
-fi
-
-if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
-    echo "stage 4: Decoding"
-    nj=16
-    if [[ $(get_yaml.py ${train_config} model-module) = *transformer* ]] || \
-       [[ $(get_yaml.py ${train_config} model-module) = *conformer* ]] || \
-       [[ $(get_yaml.py ${train_config} model-module) = *maskctc* ]] || \
-       [[ $(get_yaml.py ${train_config} etype) = transformer ]] || \
-       [[ $(get_yaml.py ${train_config} dtype) = transformer ]]; then
-        average_opts=
-        if ${use_valbest_average}; then
-            recog_model=model.val${n_average}.avg.best
-            average_opts="--log ${expdir}/results/log"
-        else
-            recog_model=model.last${n_average}.avg.best
-        fi
-        average_checkpoints.py --backend ${backend} \
-			        --snapshots ${expdir}/results/snapshot.ep.* \
-			        --out ${expdir}/results/${recog_model} \
-			        --num ${n_average} \
-                    ${average_opts}
-    fi
-
-    pids=() # initialize pids
-    for rtask in ${recog_set}; do
-    (
-        decode_dir=decode_${rtask}_$(basename ${decode_config%.*})
-        feat_recog_dir=${dumpdir}/${rtask}/delta${do_delta}
-
-        # split data
-        splitjson.py --parts ${nj} ${feat_recog_dir}/data.json
-
-        #### use CPU for decoding
-        ngpu=0
-
-        ${decode_cmd} JOB=1:${nj} ${expdir}/${decode_dir}/log/decode.JOB.log \
-            asr_recog.py \
-            --config ${decode_config} \
-            --ngpu ${ngpu} \
-            --backend ${backend} \
-            --debugmode ${debugmode} \
-            --recog-json ${feat_recog_dir}/split${nj}utt/data.JOB.json \
-            --result-label ${expdir}/${decode_dir}/data.JOB.json \
-            --model ${expdir}/results/${recog_model}
-
-        score_sclite.sh --wer true ${expdir}/${decode_dir} ${dict}
-
-    ) &
-    pids+=($!) # store background pids
-    done
-    i=0; for pid in "${pids[@]}"; do wait ${pid} || ((++i)); done
-    [ ${i} -gt 0 ] && echo "$0: ${i} background jobs are failed." && false
-    echo "Finished"
-fi
