+++ espnet/nets/pytorch_backend/ctc.py
@@ -5,91 +5,8 @@ import numpy as np
 import six
 import torch
 import torch.nn.functional as F
-import gtn
 
 from espnet.nets.pytorch_backend.nets_utils import to_device
-
-class GTNCTCLossFunction(torch.autograd.Function):
-
-    @staticmethod
-    def create_ctc_graph(target, blank_idx):
-        g_criterion = gtn.Graph(False)
-        L = len(target)
-        S = 2 * L + 1
-        for l in range(S):
-            idx = (l - 1) // 2
-            g_criterion.add_node(l == 0, l == S - 1 or l == S - 2)
-            label = target[idx] if l % 2 else blank_idx
-            g_criterion.add_arc(l, l, label)
-            if l > 0:
-                g_criterion.add_arc(l - 1, l, label)
-            if l % 2 and l > 1 and label != target[idx - 1]:
-                g_criterion.add_arc(l - 2, l, label)
-        g_criterion.arc_sort(False)
-        return g_criterion
-
-    @staticmethod
-    def forward(ctx, log_probs, targets, blank_idx=0, reduction="none"):
-        B, T, C = log_probs.shape
-        losses = [None] * B
-        scales = [None] * B
-        emissions_graphs = [None] * B
-
-        def process(b):
-            # create emission graph
-            g_emissions = gtn.linear_graph(T, C, log_probs.requires_grad)
-            cpu_data = log_probs[b].cpu().contiguous()
-            g_emissions.set_weights(cpu_data.data_ptr())
-
-            # create criterion graph
-            g_criterion = GTNCTCLossFunction.create_ctc_graph(targets[b], blank_idx)
-            # compose the graphs
-            g_loss = gtn.negate(
-                gtn.forward_score(gtn.intersect(g_emissions, g_criterion))
-            )
-
-            scale = 1.0
-            if reduction == "mean":
-                L = len(targets[b])
-                scale = 1.0 / L if L > 0 else scale
-            elif reduction != "none":
-                raise ValueError("invalid value for reduction '" + str(reduction) + "'")
-
-            # Save for backward:
-            losses[b] = g_loss
-            scales[b] = scale
-            emissions_graphs[b] = g_emissions
-
-        gtn.parallel_for(process, range(B))
-
-        ctx.auxiliary_data = (losses, scales, emissions_graphs, log_probs.shape)
-        loss = torch.tensor([losses[b].item() * scales[b] for b in range(B)])
-        return torch.mean(loss.cuda() if log_probs.is_cuda else loss)
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        losses, scales, emissions_graphs, in_shape = ctx.auxiliary_data
-        B, T, C = in_shape
-        input_grad = torch.empty((B, T, C))
-
-        def process(b):
-            gtn.backward(losses[b], False)
-            emissions = emissions_graphs[b]
-            grad = emissions.grad().weights_to_numpy()
-            input_grad[b] = torch.from_numpy(grad).view(1, T, C) * scales[b]
-
-        gtn.parallel_for(process, range(B))
-
-        if grad_output.is_cuda:
-            input_grad = input_grad.cuda()
-        input_grad *= grad_output / B
-
-        return (
-            input_grad,
-            None,  # targets
-            None,  # blank_idx
-            None,  # reduction
-        )
 
 class CTC(torch.nn.Module):
     """CTC module
@@ -124,11 +41,9 @@ class CTC(torch.nn.Module):
 
             self.ctc_loss = warp_ctc.CTCLoss(size_average=True, reduce=reduce)
         elif self.ctc_type == "gtnctc":
+            from espnet.nets.pytorch_backend.gtn_ctc import GTNCTCLossFunction
+
             self.ctc_loss = GTNCTCLossFunction.apply
-        elif self.ctc_type == "k2":
-            reduction_type = "sum" if reduce else "none"
-            from espnet.nets.pytorch_backend.ctc_graph import K2CTCLoss
-            self.ctc_loss = K2CTCLoss(odim, reduction=reduction_type)
         else:
             raise ValueError(
                 'ctc_type must be "builtin" or "warpctc": {}'.format(self.ctc_type)
@@ -153,12 +68,6 @@ class CTC(torch.nn.Module):
             targets = [t.tolist() for t in th_target]
             log_probs = torch.nn.functional.log_softmax(th_pred, dim=2)
             return self.ctc_loss(log_probs, targets, 0, "none")
-        elif self.ctc_type == "k2":
-            th_pred = th_pred.log_softmax(2)
-            loss = self.ctc_loss(th_pred, th_target, th_ilen, th_olen)
-            # Batch-size average
-            loss = loss / th_pred.size(1)
-            return loss
         else:
             raise NotImplementedError
